All operations now serialized
Because multiple operations cannot be executed simultaneously on a
connection, operations are queued and then executed in the order
specified (FIFO).

When an operation is called on the connection object by the user, it adds
itself to the queue.  If the operation is the only one on the queue, it
executes itself.  If another operation is called while the first is still
executing, it is simply added to the end of the queue and when the first
op finishes, the second is started.

A generic javascript array holds the operations.  Each operation is just a
function and its arguments.  Each operation is responsible for kicking off
the next operation by calling nextOp when it is completed.

Tests were added for serialized queries and a serialized transaction.

This will not likely be used much when MARS support is enabled.  However,
the option will likely still remain since there are operations that aren't
MARS friendly.