Row count returned in result set for INSERT, UPDATE and DELETE
When a query results in no rows (no metadata), the rowcount property
is returned instead.

This is meant for cases using INSERT, UPDATE, and DELETE that don't
return results, but return a row count instead.

In both cases the meta property is returned, and can be used to
determine if rows or rowcount is present.  meta is a null when
rowcount is returned.

When rows are returned, the rowcount isn't returned because most
of the time it's -1, which is essentially useless.

When (non forward-only) cursors are properly implemented, it is
likely that rowcount will be returned with rows, though meta may
be updated to reflect that the results are scrollable too.

To clean up the code a bit, accessors were written for ResultSet
instead of allowing direct access to the data fields.

Microsoft Internal Review Id: jaykint-0b08af06923442ef826863e67233e2e2