Multiple result sets supported
When several queries are batched together, return each query's results
separately.

If a callback is passed to the query function, the callback will be called
once for each result set.  A new flag, more, has been added to the passed
parameters to a user callback, so that a callback will know if more is coming
or not.  This can be used to close the connection if no more results are
coming.

When events are used, the events are called in succession multiple times: meta
once for each result set, row for each row, column for each column in a row.
The done event is called only once when all result sets have been processed.

TODO: Should we add a new event for a finished result set before starting the
next one?

Tests added for multiple result sets.

Fixes #16
Fixes #21

Microsoft Internal Review Id: jaykint-d1488f1ee64d4e4f85629b403cf78f54