Return the SQL Server data type and UDT classname in the metadata
Added a new field to the metadata, sqlType, that returns the datatype
as given by SQL Server (e.g., "varchar", "bigint", etc.)

When this type is "udt", an additional field, "udtType" is added that
contains the class name for the UDT type.  This can be used to see if
the type is a spatial type (e.g. "geography").

Also contain minor whitespace corrections (tabs to spaces).

Tests were updated to reflect the change.