Connections closed when an error occurs using sql.query or sql.queryRaw
When sql.query or sql.queryRaw were called and an error occurred during
processing, the connection was not closed and would stay open for several
minutes until collected.  Now the connections are closed immediately.

Also added a try/catch around a user's callback for query and queryRaw so
if an exception is thrown, the connection is closed.  The exception is
then rethrown with additional information.

The try/catch will be upgraded to use domains in the near future.

Fixes #121